GH-399 migrate SPARQLUpdateDataBlockParser to sparql parser module
